refactor: reestruturando para MVC
