+++ main.py
@@ -1,112 +1,21 @@
 '''
 Main file for the project
 '''
-from prizma_to_brapa.utils.oto_operations import get_oto, generate_oto_dict, save_new_oto
-from prizma_to_brapa.utils.validators import has_semi_vowels, has_vowels
-from prizma_to_brapa.utils.validators import has_consonants
-from prizma_to_brapa.utils.validators import can_add_r_or_l
-from prizma_to_brapa.utils.extractors import extract_from_consonant_vowel
-from prizma_to_brapa.utils.extractors import extract_from_semivowel_vowel
-from prizma_to_brapa.utils.extractors import extract_from_vowel
-from prizma_to_brapa.utils.extractors import extract_from_vowel_vowel
-from prizma_to_brapa.utils.extractors import extract_from_vowel_consonant
-from prizma_to_brapa.utils.extractors import extract_from_consonant_consonant
+# import sys
+# import traceback
+# import tkinter as tk
+# from tkinter import messagebox
+from prizma_to_brapa.views.interface import PrizmaToBrapaGUI
 
 
 def main():
     """
-    Main function for the project.
+    Entry point for the application.
 
-    Gets an OTO file, parses it and generates a new one with the corresponding
-    aliases for each line, following the rules of the Brapra project.
-
-    The aliases are generated by splitting the original alias into its parts and
-    checking if it has a vowel, a consonant, or a semivowel. Depending on this, the
-    corresponding extractor function is called to generate the new alias.
-
-    After generating all the new aliases, the new OTO file is saved with the
-    generated aliases.
-
-    The function also handles cases where the alias has a hyphen, and adds the
-    corresponding hyphen to the new alias.
+    Initializes the Tkinter event loop with an instance of MainWindow.
     """
-    oto_dataframe = get_oto()
-    new_oto_entries = []
-
-    for oto_line in oto_dataframe.itertuples():
-        alias = oto_line.alias
-        new_alias = ""
-        add_hyphen_start = False
-        add_hyphen_end = False
-        add_ending_r_or_l = False
-
-        # Check if the alias has a vowel, a semivowel, or a consonant
-        alias_check = alias
-        has_vowel, alias_check = has_vowels(alias_check)
-        has_semivowel, alias_check = has_semi_vowels(alias_check)
-        has_consonant, alias_check = has_consonants(alias_check)
-
-        # Handle cases without spaces in the alias
-        if " " not in alias:
-            if has_vowel and has_consonant:
-                new_alias = extract_from_consonant_vowel(alias)
-                add_hyphen_start = True
-
-            elif has_vowel and not has_consonant:
-                if has_semivowel:
-                    new_alias = extract_from_semivowel_vowel(alias)
-                    add_hyphen_start = True
-                else:
-                    new_alias = extract_from_vowel(alias)
-
-            elif not has_vowel and has_consonant:
-                new_alias = extract_from_consonant_consonant(alias)
-
-            else:
-                raise Exception(f"Invalid alias: {oto_line.alias}")
-
-        # Handle cases with spaces in the alias
-        else:
-            alias_parts = alias.split()
-            if len(alias_parts) == 2:
-                if has_vowel and has_consonant:
-                    new_alias = extract_from_vowel_consonant(
-                        vogal=alias_parts[0], cons=alias_parts[1])
-                    add_hyphen_end = True
-                    add_ending_r_or_l = can_add_r_or_l(alias_parts[1])
-
-                elif has_vowel and not has_consonant:
-                    new_alias = extract_from_vowel_vowel(alias)
-
-                    if has_semivowel:
-                        add_hyphen_end = True
-
-                elif not has_vowel and has_consonant:
-                    pass
-                else:
-                    raise Exception(f"Invalid alias: {oto_line.alias}")
-
-        # Add the new alias to the list and add the corresponding hyphen
-        if new_alias:
-            oto_dict = generate_oto_dict(new_alias, oto_line)
-            new_oto_entries.append(oto_dict)
-
-            if add_hyphen_start:
-                oto_dict = generate_oto_dict("-" + new_alias, oto_line)
-                new_oto_entries.append(oto_dict)
-
-            if add_hyphen_end:
-                oto_dict = generate_oto_dict(new_alias + "-", oto_line)
-                new_oto_entries.append(oto_dict)
-
-            if add_ending_r_or_l:
-                for suffix in ["r", "r-", "l", "l-"]:
-                    oto_dict = generate_oto_dict(new_alias + suffix, oto_line)
-                    new_oto_entries.append(oto_dict)
-
-    # Save the new OTO file
-    save_new_oto(oto=new_oto_entries)
-    print("New OTO file generated.")
+    app = PrizmaToBrapaGUI()
+    app.mainloop()
 
 
 if __name__ == "__main__":
